numbers type change

+++ calculator.py
@@ -7,8 +7,8 @@ inp = input()
 
 def main(inp):
     string_to_list = inp.split(" ")
-    a = int(string_to_list[0])
-    b = int(string_to_list[2])
+    a = float(string_to_list[0])
+    b = float(string_to_list[2])
     operator = string_to_list[1]
     return a, b, operator
